+++ scripts/add_brownfield.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Contributors to PyPSA-Eur <https://github.com/pypsa/pypsa-eur>
+# SPDX-FileCopyrightText: Open Energy Transition gGmbH and contributors to PyPSA-Eur <https://github.com/pypsa/pypsa-eur>
 #
 # SPDX-License-Identifier: MIT
 """
@@ -206,7 +206,9 @@ def disable_grid_expansion_if_limit_hit(n):
                 n.global_constraints.drop(name, inplace=True)
 
 
-def adjust_renewable_profiles(n, input_profiles, params, year):
+def adjust_renewable_profiles(
+    n, input_profiles, params, year, tyndp_renewable_carriers
+):
     """
     Adjusts renewable profiles according to the renewable technology specified,
     using the latest year below or equal to the selected year.
@@ -218,7 +220,12 @@ def adjust_renewable_profiles(n, input_profiles, params, year):
         pd.Series(dr, index=dr).where(lambda x: x.isin(n.snapshots), pd.NA).ffill()
     )
 
-    for carrier in params["carriers"]:
+    # TODO: hotfix remove filter for tyndp_renewable_carriers after tyndp generators are added
+    if len(tyndp_renewable_carriers) > 0:
+        logger.info(
+            f"Hotfix until TYNDP renewable carriers are added. Skipping renewable carriers '{', '.join(tyndp_renewable_carriers)}'."
+        )
+    for carrier in set(params["carriers"]) - set(tyndp_renewable_carriers):
         if carrier == "hydro":
             continue
 
@@ -349,7 +356,11 @@ if __name__ == "__main__":
 
     n = pypsa.Network(snakemake.input.network)
 
-    adjust_renewable_profiles(n, snakemake.input, snakemake.params, year)
+    tyndp_renewable_carriers = snakemake.params.tyndp_renewable_carriers
+
+    adjust_renewable_profiles(
+        n, snakemake.input, snakemake.params, year, tyndp_renewable_carriers
+    )
 
     add_build_year_to_new_assets(n, year)